+++ opensubdiv/src/osd/cuda_vertex_buffer.rs
@@ -17,7 +17,7 @@ impl Drop for CudaVertexBuffer {
 
 impl CudaVertexBuffer {
     #[inline]
-    pub fn new(elements_len: u32, vertices_len: u32) -> CudaVertexBuffer {
+    pub fn new(elements_len: usize, vertices_len: usize) -> CudaVertexBuffer {
         let ptr = unsafe {
             sys::osd::CudaVertexBuffer_Create(
                 elements_len.try_into().unwrap(),
@@ -34,13 +34,13 @@ impl CudaVertexBuffer {
 
     /// Returns how many elements defined in this vertex buffer.
     #[inline]
-    pub fn elements_len(&self) -> u32 {
+    pub fn elements_len(&self) -> usize {
         unsafe { sys::osd::CudaVertexBuffer_GetNumElements(self.0) as _ }
     }
 
     /// Returns how many vertices allocated in this vertex buffer.
     #[inline]
-    pub fn vertices_len(&self) -> u32 {
+    pub fn vertices_len(&self) -> usize {
         unsafe { sys::osd::CudaVertexBuffer_GetNumVertices(self.0) as _ }
     }
 
@@ -66,13 +66,13 @@ impl CudaVertexBuffer {
     pub fn update_data(
         &mut self,
         src: &[f32],
-        start_vertex: u32,
-        vertices_len: u32,
+        start_vertex: usize,
+        vertices_len: usize,
     ) {
         // do some basic error checking
         let elements_len = self.elements_len();
 
-        if (start_vertex * elements_len) as usize > src.len() {
+        if start_vertex * elements_len > src.len() {
             panic!(
                 "Start vertex is out of range of the src slice: {} ({})",
                 start_vertex,
@@ -80,7 +80,7 @@ impl CudaVertexBuffer {
             );
         }
 
-        if (vertices_len * elements_len) as usize > src.len() {
+        if vertices_len * elements_len > src.len() {
             panic!(
                 "vertices_len is out of range of the src slice: {} ({})",
                 vertices_len,
